.claude/worktrees/focused-curie/desktop-app/requirements.txt
customtkinter>=5.2.0
tkcalendar>=1.6.1
playwright>=1.40.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
openpyxl>=3.1.0
psutil>=5.9.0
pywin32>=306
pywinauto>=0.6.8
